requirements.txt
pandas==2.2.3
requests==2.32.4
azure-storage-blob==12.19.1
